requirements.txt
flask
gunicorn
boto3
pillow
python-dotenv

